[kotlin] Make KotlinUnusedImportInspection run on every change in the file
If a new import added, any PSI element in the rest of
the file might be highlighted as "Redundant qualifier". Same goes the other way round -
after removing any call in any place of the
file, some import might become unused or
used. That's why this inspection should be
run on every change in the file.

However, the problem with false
positives/negatives can still be present,
because at the moment K2 frontend caches
FIR declarations for the bodies, and it
may cause problems (see KTIJ-25749 for
the details)

GitOrigin-RevId: 848e3da7053b6e4b6db36c3bba85bfdb697b16ff