KTIJ-26224 [kotlin][scripting]: highlighting doesn't work after project reopening
Script configuration might not be available (loaded, applied) at the moment
when a decision about highlighting is made. Even worse, the decision is
cached. Configuration loading/availability/caching is a design level issue
to be solved separately.
This commit updates the logic of "should highlight script" predicate.
It removes the checks associated with script configuration availability,
thus raising the probability of successful highlighting.

^KTIJ-26224 fixed

GitOrigin-RevId: c6a69c752896fa6e7bc77fbef70e0429f165f0c2