[debugger] IDEA-320218 Fix breakpoints for crossinline lambdas
Crossinline lambdas are compiled into a class with a name different from the usual lambda class name. As this generated name is hard to recover exactly by ClassNameCalculator, this commit adds a heuristic approach for choosing the correct lambda. Namely, when there is only one lambda on the line (or only one with the required name), we can take it. Alternatively, we try to guess lambda sequential number by searching the whole method.


Merge-request: IJ-MR-108392
Merged-by: Maksim Zuev <Maksim.Zuev@jetbrains.com>

GitOrigin-RevId: faf10128700692d4f94bf1656049a39123a885e5