[kotlin] KTIJ-26308 Script definition might be unavailable during resolve
The root cause of the issue is a change made for KT-60193.
Previously, the compiler used so-called default definition in cases
where the expected one was not available [1]. Default definition couldn't provide
enough information for the resolve, and the latter could fail without a clear reason.

This commit is aimed to provide the compiler with a definition in the form of
deferred one - synchronous wrapper polling a cache for actual value [2]. Such an
approach is practical for the case where definition is still loaded by the moment
of its need. Real context is one of the Platform highlighting paths [3] causing
resolve.

------------------------------------------------------------------------------
[1]: org.jetbrains.kotlin.idea.core.script.ScriptDefinitionsManager.findDefinition
[2]: org.jetbrains.kotlin.idea.core.script.DeferredScriptDefinition
[3]: com.intellij.codeInsight.daemon.impl.IdentifierHighlighterPass


Merge-request: IJ-MR-111994
Merged-by: Andrei Klunnyi <andrei.klunnyi@jetbrains.com>

GitOrigin-RevId: 51120fdec2a66be81c00053c8104baddda816747