[kotlin] J2K: fix exposed visibility errors
In case of exposed visibility errors, there are two general directions:

1. Make the used declaration more visible
2. Make the call-site declaration less visible

It seems that the first fix is more useful. If we restrict visibility of a use-site declaration
it may just cause new errors in other places.

Probably the most popular case in practice is exposing visibility of a
package-private (internal in Kotlin) declaration by a public declaration.
In this case it seems reasonable to make the internal declaration public.

^KTIJ-9791 Fixed
^KTIJ-5025 Fixed

GitOrigin-RevId: 0592d12fcc53593df75c081e3d5520725fb206ea