[kotlin][uast] drop resolve from getImplicitReturn
This is not fully correct in the case of lambda with `Unit` return type
and non-`Unit` return type of the last statement:
```kotlin
fun foo() {
  42.apply {
    "str"
  }
}
```

Because here `apply` has `Unit` return type,
so we shouldn't have the implicit return here, but we will create it anyway.

So effectively, this code will mean:
```kotlin
fun foo() {
  42.apply {
    return@apply "str"
  }
}
```
in terms of UAST what is potentially wrong in some corner cases,
but we agree with this behavior because such real type checks are too expensive.

But it is correct in the case of [Unit] as a return type of the last statement
```kotlin
fun foo() {
    42.apply {
        return@apply println(this)
    }
}
```

^KTIJ-26541

GitOrigin-RevId: 4265c9112339200297b8690c0ab30ba30ef2816c