[kotlin] Fix 'ResolutionFacade' provision in 'IdeFe10AnalysisFacade'
The previous implementation delegated straight to 'KotlinCacheService.getResolutionFacade(List<KtElement>)' that has strict logic for handling files outside the module scope. In particular, it throws an exception when it finds several candidate modules (see 'createFacadeForFilesWithSpecialModuleInfo()').

The new implementation chooses the first available special file. It's obviously less correct in the formal sense, however such an approach is often used in K1 code in places that deals with potentially unrelated files (as during Bytecode Tool Window compilation, or in the evaluator).

GitOrigin-RevId: 4b1525531bc1ac2db79ace3daccec3c6810276df