[kotlin] IR / Debugger: “Exception during IR lowering” occurs periodically during evaluating local function
CodeFragmentUtils.USED_FOR_COMPILATION_IN_IR_EVALUATOR is added to identify the code fragment used for compilation in the IR evaluator. It is necessary to ignore caching in the analysis (PerFileAnalysisCache.getAnalysisResults) as it can cause problems related to data inconsistency. However, it's important not to ignore the code fragments that aren't used for compilation (e.g for highlighting), as it can lead to a decrease in the evaluator's performance.

^IDEA-326794

GitOrigin-RevId: a2d945fa222044f46605ec8044bddd5c102288cd