[Kotlin] Don't suggest "Copy annotation to actual" fix for f/o and typealias
Don't suggest quick-fix "Copy annotation from expect to actual" for
fake overrides and `actual typealias`, even if the source is available.

This is detected by full fqName of method or class.
If it is the same as the name of `expect` declaration, we are sure
that actual declaration is inside `actual class`.

^KTIJ-26633

GitOrigin-RevId: 21f2b7cb007f539b3275989e86d0ac5a04bb98b4