[kotlin] fix `kotlinc.kotlin-stdlib` library
Problem: `kotlinc.kotlin-stdlib` is treated as a real stdlib by
`KotlinStdlibCache` caches that leads to problems with Fe10 descriptors
Solution: leave `kotlin-stdlib-jdk8` as it was before to avoid such problems.
Note: `kotlin-stdlib-jdk8` is an empty artifact, but we need it to avoid resolution
problems. The proper fix should be done probably on `KotlinStdlibCache` side

^KTIJ-26637 Fixed

GitOrigin-RevId: 56d68ddeb1492ecc6759678ed0ae64caa66d0967