KTIJ-25334 Gradle 8.1: Unresolved references in IDE for build.gradle.kts
The essence of the problem was that the compiler frontend worked in compatibility
mode for the "new" inference. In the case of Kotlin-Java interop, a less specific
overload candidate was preferred to the one having SAM with receiver argument.

Gradle embedded Kotlin compiler is supplied with an option which was missing in
IDEA's one: `-XXLanguage:+DisableCompatibilityModeForNewInference`. Due to this fact,
Gradle worked fine in CLI mode whereas IDEA didn't.

Arguments for scripts are specified on a level of their definitions/configurations.
IDEA retrieves them and passes to the compiler. Gradle 8.x versions have bug -
arguments are not available.

This commit tracks the absence of the compiler arguments for Gradle scripts
and adds them on IDE level.

^KTIJ-25334 fixed

GitOrigin-RevId: 4fb41385a1ed1aaf49cb36926d3e7f1a8e1916ad