Fix search scope for `KtLibrarySourceModules` in scope provider.
Without this fix, when creating a search scope for a library,
the contents of that library itself was not included in the search
scope.

That causes issue for the use in evaluate expression. When
performing evaluate expression at a breakpoint in a binary
library dependency, class files for the library itself
could not be found.

^KTIJ-26694 Fixed

GitOrigin-RevId: fcc79d0d25c29bb76824ab17176bf6b0633aa2e7