KTIJ-25236 kotlin.Cloneable is unavailable for build.gradle.kts
Built-in classes like `kotlin.Cloneable` require sdk and std-lib to be instantiated.
In this particular case, the error happened due to the missing sdk.

All sdk-s for a project are cached in `SdkInfoCache` utilizing `ModuleInfo` as a key.
In the context of scripts, these are `ScriptModuleInfo` and `ScriptDependenciesInfo`.

Sdk for a script is taken from its compilation configuration. So missing configuration
means the absence of sdk for corresponding `ModuleInfo`(s). This fact is reflected in
`SdkInfoCache` as a mapping to `null`. This happens, for example, if a module is not
included in Gradle build - its nested script configurations are not available. It's a
valid case now.

After the first opened script without configuration left its "null sdk" trace in the
cache, all subsequently opened scripts got the same "broken" sdk value. This happened
because `equals`/`hashCode` methods defined for `ScriptModuleInfo` and
`ScriptDependenciesInfo` didn't take a script path into account.

^KTIJ-25236 fixed

GitOrigin-RevId: b9fe94f5b5d38c193002f35ab09ccd9eceb5fc24