[kotlin] J2K: try to preserve all line breaks as accurately as possible
Now, instead of a boolean flag "has line breaks before/after",
we save the actual number of line breaks for each element and print them.

Usually, we only print line breaks after an element, and ignore line breaks before,
because otherwise we would be printing line breaks twice for sibling elements.
But for some elements (for example, arguments), we save and print line break before,
because it is easier from an implementation point of view.

The final result is affected by the Kotlin code formatter (for example,
if there are too many line breaks in a row, the formatter may delete some,
according to the project code style).

Note: the whole J2K formatting scheme is not perfect,
because instead of saving whitespace and comments as separate tree elements,
we attach them to "real" elements as attributes. But some elements may be
replaced by synthetic ones, or changed by a processing and the formatting
will not be accurately preserved. We try to avoid such cases though.

^KTIJ-18457 Fixed

GitOrigin-RevId: 318af33c019fbd0ccb887c426bb70948a738f2a2