[kotlin][debugger] IDEA-320283 fix evaluation in presence of delegated properties with inline accessors
One of the compiler lowerings, namely `PropertyReferenceLowering`,
optimizes usages of property references in some cases,
and if a containing delegated property accessor is inline,
it might need this accessor's bytecode.

If an accessor function is defined in a different module,
IDE tries to acquire its bytecode via the index, however,
the index doesn't cover classfiles from compiler output,
and that leads to errors like KTIJ-26706 and IDEA-320283.

One of the ways to ensure that the bytecode exists is to include files
that contain inline accessors in the compilation process. For example,
if file f1.kt from the analysis result contains property `val p by D()`,
and `D.getValue` is an inline function defined in file f2.kt,
then f2.kt must be present in the analysis result.
If f2.kt contains another property `val q by E()`,
and `E.getValue` is an inline function defined in f3.kt,
then f3.kt must be present in the analysis result. And so on.

Compiler-side part of the fix: 58350b83a8ac28c01f71175c2c30240bcd9bb464


Merge-request: KT-MR-12012
Merged-by: Alexander Kuznetsov <Aleksander.Kuznetsov@jetbrains.com>

GitOrigin-RevId: af7e2c409121c047e093b65e41bd1dc6b586fd50