[kotlin][debugger] IDEA-331853 don't perform dumb check at the start of evaluation
First, it doesn't save us from getting an INRE when an IDE enters dumb mode after this check (INREs in evaluator contribute to EA-404968 and happen in `InlineFunctionAnalyzer`).

Second, in order to fix evaluator cases of that very EA-404968, we now catch an INRE coming from anywhere during the analysis+compilation, and wrap it in EvaluationException. Basically, our evaluation code is dumb-safe without this check at the beginning.

Third, it doesn't allow an IDE evaluate already compiled code fragments, like conditional breakpoints. If an IDE enters dumb mode during an active debugging session (which isn't an unlikely situation), it will spawn a modal window each time execution reaches a line with a conditional breakpoint, whereas it is able to evaluate this expression without using indexes.

It's also possible to just move dumb check past the point of obtaining a cached code fragment, but there would be little point in it, as the first two observations remain intact.

GitOrigin-RevId: 66b6224a27b7a0dd88d202265a65042d1b30ed2a