[kotlin][debugger] IDEA-331853 improve code fragment caching -- invalidate it only on debug refresh
Current implementation invalidates cache on every PSI modification. Whereas it's a good approach to ensure cached elements won't get stale, it poses a problem in the dumb mode, as entering the dumb mode counts as a PSI modification.

The problem is that a debugger can't recompile expressions in the dumb mode, so a user gets errors in the UI, like a modal window with failed conditional breakpoint, which can be sudden and annoying. The interesting part is that, if it weren't for the cache invalidation, the debugger would still be able to evaluate this expression. So we should find another way to invalidate caches.

A cache for code fragments can be represented as a map `context -> text of code -> compiled version`, where `context` specifies the place in code the code fragment is calculated in, and `text of code` is simply a written code. So, even if the text is changed, there won't be an erroneous cache hit, and we're safe.

In some cases a cache can get stale, even if its text or context wasn't changed, for example, after the HotSwap or re-debugging, when the identifiers used in an expression may start leading to an entirely different declaration. In such cases, we need to invalidate the caches. This commit proposes doing invalidation when the session has to be "refreshed" (after reloading classes, setting custom values in Variables view, etc.), and when a debugger session is attached/detached.


Merge-request: IJ-MR-114735
Merged-by: Alexander Kuznetsov <Aleksander.Kuznetsov@jetbrains.com>

GitOrigin-RevId: ccdd4b7fa995b88bbe1d02850796c8265e3b671d