[kotlin] J2K: ConvertGettersAndSettersToPropertyProcessing cleanup
Rearrange code, simplify a bit and add some tests to assert existing behavior.

AFAIK behavior is changed only in one corner case (see test `InInterface.java`):

Now, if getter and setter of an interface have different types,
the getter will be converted to a property, like it was done for a usual class.
I couldn't figure out why an interface has to be treated specially here
and decided to unify behavior.

GitOrigin-RevId: d8d7223970c3367c39566dab6198d8c61497be8f